Backed out 4 changesets (bug 1766377) for causing mochitest failures on HyperTextAccessible.cpp CLOSED TREE
Backed out changeset 6335511477a4 (bug 1766377)
Backed out changeset e8108d6fab9f (bug 1766377)
Backed out changeset 80d16d55c490 (bug 1766377)
Backed out changeset 1a6583d6e52e (bug 1766377)

+++ accessible/windows/msaa/Compatibility.cpp
@@ -117,8 +117,7 @@ static DWORD WINAPI InSendMessageExHook(LPVOID lpReserved) {
 static LONG CALLBACK
 DetectInSendMessageExCompat(PEXCEPTION_POINTERS aExceptionInfo) {
   DWORD exceptionCode = aExceptionInfo->ExceptionRecord->ExceptionCode;
-  if (exceptionCode == static_cast<DWORD>(RPC_E_CANTCALLOUT_ININPUTSYNCCALL) &&
-      NS_IsMainThread()) {
+  if (exceptionCode == RPC_E_CANTCALLOUT_ININPUTSYNCCALL && NS_IsMainThread()) {
     sInSendMessageExHackEnabled = true;
     // We don't need this exception handler anymore, so remove it
     if (RemoveVectoredExceptionHandler(sVectoredExceptionHandler)) {
